[tests] Remove permanent skipped tests and move manual benchmarks out of pytest
Addresses the S4/S5 findings from the dead-test audit. Three groups:

A. Delete dead unconditionally-skipped tests:
   - test_attention.py::test_causal_mask_blocking (used removed
     AttentionMask.blocked()/.max_Pos/.KeyPos)
   - test_audio.py::test_resolve_audio_pointer (gone external mp3)
   - test_gdn_kernels.py::test_kernels_match_hf_without_l2norm (skipped on
     line 1; asserted behavior the kernels intentionally do not satisfy)

B. Move the TPU+Llama Arrow Flight benchmarks out of pytest into a runnable
   script (marin.rl.scripts.benchmark_weight_transfer); they require real TPU
   and HuggingFace access and were never collected as tests anyway.

C. Delete the remaining permanently-skipped tests per maintainer direction:
   whisper_test.py, test_train_asr.py, test_hf_gpt2_serialize.py::
   test_mistral_gpt2_roundtrip, test_audio.py::test_hf_audio_cache_pipeline.

Conditional dependency/backend/environment guards (skip_if_no_soundlibs,
skip_if_no_torch, skip_if_hf_model_not_accessible, device/TPU/GCS checks) are
left intact.

Fixes #6425
