+++ lib/levanter/tests/test_attention.py
@@ -61,28 +61,6 @@ def _make_explicit_abstract_mesh() -> AbstractMesh:
     )
 
 
-@pytest.mark.skip
-def test_causal_mask_blocking():
-    pos = hax.Axis("pos", 128)
-    key_pos = pos.alias("key_pos")
-
-    mask = AttentionMask.causal()
-
-    blocked_mask = mask.blocked(pos, 16).blocked(key_pos, 16)
-    assert blocked_mask.max_Pos.size == 128 // 16
-    assert blocked_mask.KeyPos.size == 128 // 16
-
-    mat_blocked = blocked_mask.materialize()
-
-    assert hax.all(mat_blocked == hax.nn.attention.causal_mask(pos.resize(8), key_pos.resize(8)))
-
-    mat_mask = mask.materialize()
-
-    for i in range(8):
-        for j in range(8):
-            assert mat_blocked.array[i, j] == jnp.any(mat_mask.array[i * 16 : (i + 1) * 16, j * 16 : (j + 1) * 16])
-
-
 def test_causal_mask_slicing():
     pos = hax.Axis("pos", 128)
     key_pos = pos.alias("key_pos")
