+++ lib/levanter/tests/test_gdn_kernels.py
@@ -469,58 +469,6 @@ def test_extreme_gates_no_nans_and_parity():
 
 
 @skip_if_no_torch
-def test_kernels_match_hf_without_l2norm():
-    # TODO: fix edge case? although per original paper L2 norm is needed for stability
-    pytest.skip("not matching HF implementation")
-
-    import torch  # noqa: PLC0415  # optional dep: torch
-
-    hf_chunk, hf_recur = _get_hf_kernels()
-
-    key = jax.random.PRNGKey(0)
-    B, H, L, dk, dv = 2, 3, 57, 16, 8
-
-    q, k, v, g, beta = _named_kernels_inputs(B, H, L, dk, dv, key)
-
-    # Haliax kernels with use_qk_l2norm_in_kernel=False
-    out_chunk_j, _ = chunk_gated_delta_rule(
-        q, k, v, g, beta, chunk_size=32, output_final_state=False, use_qk_l2norm_in_kernel=False
-    )
-    out_recur_j, _ = recurrent_gated_delta_rule(
-        q, k, v, g, beta, output_final_state=False, use_qk_l2norm_in_kernel=False
-    )
-
-    # HF fallback expects (B, L, H, dim) on input and transposes internally; don't move axes.
-    def to_t(arr: jnp.ndarray):
-        return torch.from_numpy(np.array(arr))
-
-    out_chunk_t, _ = hf_chunk(
-        to_t(q.array),
-        to_t(k.array),
-        to_t(v.array),
-        to_t(g.array),
-        to_t(beta.array),
-        chunk_size=32,
-        initial_state=None,
-        output_final_state=False,
-        use_qk_l2norm_in_kernel=False,
-    )
-    out_recur_t, _ = hf_recur(
-        to_t(q.array),
-        to_t(k.array),
-        to_t(v.array),
-        to_t(g.array),
-        to_t(beta.array),
-        initial_state=None,
-        output_final_state=False,
-        use_qk_l2norm_in_kernel=False,
-    )
-
-    np.testing.assert_allclose(np.array(out_chunk_j.array), _to_np(out_chunk_t), rtol=1e-4, atol=1e-4)
-    np.testing.assert_allclose(np.array(out_recur_j.array), _to_np(out_recur_t), rtol=1e-4, atol=1e-4)
-
-
-@skip_if_no_torch
 def test_recurrent_backward_matches_hf():
     """
     JAX vs HF fallback gradient parity for the recurrent (decode) kernel.
